telemetry: enforce `result` and `reason` properties for emitted metrics (#4103)
* telemetry: enforce `result` and `reason` properties for emitted metrics

Problem: Different integrations within the toolkit publish metrics differently. The 'result' and 'reason' fields are meant
to be standardized, but some integrations are adding their own fields.
Solution: Verify that any metrics emitted using `emit()` contain a result property. If the result property is 'Failed', then
a reason property must also exists. Any tests that violate this code are failed. We do not want to fail in general because
it should not create an exception during production toolkit usage.

* refactor: telemetry property check no longer uses spy

* Update src/test/globalSetup.test.ts

Co-authored-by: Justin M. Keyes <jmkeyes@amazon.com>

* Update src/test/globalSetup.test.ts

Co-authored-by: Justin M. Keyes <jmkeyes@amazon.com>

* refactor: move telemetry validation out of tests

* Update spans.test.ts

* Update TESTPLAN.md

* Apply suggestions from code review

Co-authored-by: Justin M. Keyes <jmkeyes@amazon.com>

* finish implementation of PR suggestions

* update metrics exempt list

* fix: pass validation for metric 'aws_modifySetting'

---------

Co-authored-by: Justin M. Keyes <jmkeyes@amazon.com>