fix(codecatalyst): no longer prompt builder ID sign in when opening dev env
Problem: Opening a dev env inside of the IDE will trigger logic to default to a builder ID sign in. This is because the function
will not receive the proper information when opened from menu options (the connection variable is not defined when called this
way),
triggering the default path. Opening a dev env from the
CodeCatalyst website will open it normally, because it passes the necessary information to the function that will not cause the
default behavior to activate.

Solution: Check for valid connection state instead of relying on what is passed into the function. Also, remove Builder ID login
as default, instead just route to the connections page. Although, this codepath should only run if something funky is going on.