fix: sam python tests failing due to race condition
We would run continueDebugger() to "click" the continue button
when debugging, but this failed due to there being no breakpoints
and the debugger never pausing. When trying to continue the call would
throw an error since it wasn't possible.

The spirit of the test was to click the continue button to ensure the
debugger completed. Since we do not pause anywhere, the solution this commit
does is to instead wait for the debug session to return nothing, assuming this indicates it is done.

Signed-off-by: nkomonen <nkomonen@amazon.com>