fix(codecatalyst): default to builder ID sign in if URI contains no IdC params (#4174)
Problem: If a user opens a vscode URI to open a dev environment and there are no IdC params, then the user is pointed to the login page instead of being routed to builder ID
sign in. This means the user has to click the link again after signing in.
Solution: Always default to builder ID if there are no IdC params in the URI, rather than prompt for sign in.