SAM: fix failing SAM tests (#4170)
* test: Bump sam version in CI to see if it passes now

- remove the old tech debt test to signal to us to check if a new version works. We do not need this anymore now.
- Always download the latest version of sam cli in CI instead of using the fixed version. This is back to what it was before we hard set a version.

Signed-off-by: nkomonen <nkomonen@amazon.com>

* remove node14 from sam tests

This is indicated as deprecated

Signed-off-by: nkomonen <nkomonen@amazon.com>

* fix: sam python tests failing due to race condition

We would run continueDebugger() to "click" the continue button
when debugging, but this failed due to there being no breakpoints
and the debugger never pausing. When trying to continue the call would
throw an error since it wasn't possible.

The spirit of the test was to click the continue button to ensure the
debugger completed. Since we do not pause anywhere, the solution this commit
does is to instead wait for the debug session to return nothing, assuming this indicates it is done.

Signed-off-by: nkomonen <nkomonen@amazon.com>

* install older version of python extension so CI passes

Signed-off-by: nkomonen <nkomonen@amazon.com>

---------

Signed-off-by: nkomonen <nkomonen@amazon.com>