fix: "/dev" fails with "Conversation id must exist" #4188
Problem:
"/dev" inside Q fails when a new conversation is started:
```
Sorry, we encountered a problem when processing your request. Amazon Q feature development request failed: Conversation id must exist before continuing
``` 

Retrying the operation is a workaround for this, and will make the conversation work as expected.

Solution:
Add `await` when creating the conversation. The missing `await` caused the error because the toolkit tried reading the `conversationId` before it was populated.