fix(codewhisperer): incorrect type cast if error is not AwsError #4193
Problem:
In CW error handling codepath, a throwable is directly cast to string
type, which doesn't contain includes member function. #4190

Thus it will cause a NPE when the finally block is executed where there
is a include method call:
https://github.com/aws/aws-toolkit-vscode/blob/24cd83683bd8366485eb60e7be49acae735ec89d/src/codewhisperer/service/recommendationHandler.ts#L328-L329

Solution:
Return the string form of an `Error` instead of the `Error` itself.