ci: fail on "rejected promise not handled"
Problem:
When nodejs reports "rejected promise not handled", CI does not fail.
The CI test logs must be manually inspected to notice these problems.
Requiring `await` #1638 will help, but still misses some cases.

Solution:
Fail if "rejected promise not handled" is found in the test output.
This is done in the CI script because the messages are from nodejs and
I haven't found a way to scan the nodejs output from our test runner.

Example:

    rejected promise not handled within 1 second: AssertionError [ERR_ASSERTION]: Expected values to be strictly equal:
    rejected promise not handled within 1 second: Error: Download code bindings cancelled
    rejected promise not handled within 1 second: Error: EACCES: permission denied, mkdir '/my'
    rejected promise not handled within 1 second: Error: Error stopping job
    rejected promise not handled within 1 second: Error: No access

    ERROR: Found 9 "rejected promise not handled" in test output (see above).
           This typically indicates a bug. Read https://developer.mozilla.org/docs/Web/JavaScript/Guide/Using_promises#error_handling