AmazonQ: Incorrect range fix (#4189)
* AmazonQ: Incorrect range fix

Problem

There is an issue with the range trimming logic when a customer's selection is limited to a single line. The system does not handle this case correctly and trims the selection inappropriately. This results in invalid range trimming that needs to be addressed.
An invalid range parameter results in 4xx client errors.

Solution

The algorithm was modified and the changes were validated through the addition of new unit tests.

* Update .changes/next-release/Bug Fix-219eee10-03c2-4122-9bf8-30f0c4ce6f88.json

Co-authored-by: Justin M. Keyes <jmkeyes@amazon.com>

---------

Co-authored-by: Maxim Hayes <149123719+hayemaxi@users.noreply.github.com>
Co-authored-by: Justin M. Keyes <jmkeyes@amazon.com>