feature(browser): enable basic global commands to be shared between browser/non-browser Adds functionality for sharing some basic global commands between browser and desktop versions of the toolkit. Included commands:
- View Toolkit Documentation
- View Source on Github
- About Toolkit
- Report Toolkit Issue
- List Commands (*shared but inaccessible in browser at the moment)

Not included yet:
- View Quick Start

Also removes the use of ExtContext type where used with these commands. Instead, refactor code to only pass what was used in this, which
is just the actual vscode extension context.

Send Feedback... was disabled in the browser because vue is not yet supported and this feature may not be necessary for an initial
browser version of the toolkit.