feat(commands): disable 'View Quick Start' command for non-cloud9 instances
Problem: 'View Quick Start` command is redundant- it is the same as the marketplace page users see when they install the plugin.

Solution: Remove this command from appearing in non-Cloud9 instances. Users who use Cloud9 would generally not have installed the toolkit
themselves, so displaying a guide for the provided plugin may be helpful.