tests: explicitly set TZ for local tests
In #4297 we explicitly set the TZ in a part of the
code that CI uses, but when running the unit tests locally
it does not trigger that part of the code.

Due to this the tests still fail when run locally.

Solution:

Add the TZ assignment to a part of the code when unit tests
are running locally. Now unit tests will pass when run locally

Signed-off-by: nkomonen <nkomonen@amazon.com>