refactor(gumby): use only diff.patch to apply patch #4306
Problem:
The current implementation of the diff viewer relies on the full changed source code being returned by the ExportResultArchive API. This is unnecessary and led to some customer issues when the download size was too large. We should be able to show the diff viewer using only the diff.patch file returned by the API.

Solution:
Get rid of the "parse-diff" library and instead use "diff" (pre-existing dependency) to apply the patch to the user's project.