browser: Initialize Auth in browser
This runs the Auth activation method to set it
up.

This will run without failure, but once we try to
set up auth it will break due to needing to fix the
way we handle filesystem paths when in the browser.

But we now have auth at least loading for the most part

Signed-off-by: nkomonen <nkomonen@amazon.com>