browser: support homedir when in browser
When in the browser we do not have access to the users
file system. Instead we must use the vscode provided
uris. So we use globalStorageUri as the homedir path.

Additionally, there is a racecondition to when globals.context is defined
versus accessed by defaultCacheDir, so we converted it to a function
so it can be accessed lazily and globals.context should be available
at that point

Signed-off-by: nkomonen <nkomonen@amazon.com>