browser: Run Auth Activation in browser (#4317)
* refactor: minor code reorganization

Signed-off-by: nkomonen <nkomonen@amazon.com>

* make fs not execute on import

Make modeMap a getter so that it only executes fs.constants when
actually called an not on the class instantiation.

We are doing this since it breaks in the browser

Signed-off-by: nkomonen <nkomonen@amazon.com>

* make AWS.Token actually exist

AWS.Token did not exist when running in the browser.

As a solution this is a quick hack to get it working, see
the docstring.

Signed-off-by: nkomonen <nkomonen@amazon.com>

* browser: Initialize Auth in browser

This runs the Auth activation method to set it
up.

This will run without failure, but once we try to
set up auth it will break due to needing to fix the
way we handle filesystem paths when in the browser.

But we now have auth at least loading for the most part

Signed-off-by: nkomonen <nkomonen@amazon.com>

* refactor: convert path better in SystemUtilities

Problem:

In multiple methods in SystemUtilities the given path is converted
to a URI if it is a string, but it does not respect the scheme of
a string path and parses it as a file. This results in an incorrect
scheme on the URI

Solution:

Create a new private method toUri() which converts a string path
to a URI with the appropriate scheme, or just returns the URI if
it is already a URI.

Signed-off-by: nkomonen <nkomonen@amazon.com>

* browser: debugging the browser changes

- Disable CORS when using `npm run runInBrowser` so that auth (OIDC) will
  work
- Update documentation regarding testing/debugging/running the extension
  in browser mode

Signed-off-by: nkomonen <nkomonen@amazon.com>

* browser: dont check icon paths when in browser

Now if in browser we will not check if the paths exist for
dark and light theme.

Also we needed to maintain the sync of this function so we
do not have a good way to do this even if the URI exists in the
browser

Signed-off-by: nkomonen <nkomonen@amazon.com>

* browser: support homedir when in browser

When in the browser we do not have access to the users
file system. Instead we must use the vscode provided
uris. So we use globalStorageUri as the homedir path.

Additionally, there is a racecondition to when globals.context is defined
versus accessed by defaultCacheDir, so we converted it to a function
so it can be accessed lazily and globals.context should be available
at that point

Signed-off-by: nkomonen <nkomonen@amazon.com>

* browser: debugging in chrome works!

Signed-off-by: nkomonen <nkomonen@amazon.com>

* rename: browser script

Signed-off-by: nkomonen <nkomonen@amazon.com>

* rename: browserRun script package.json

Signed-off-by: nkomonen <nkomonen@amazon.com>

* browser: run chrome window w/ debugging

- Now we can run a Chrome window with the extension
in it and debug from VS Code
- Added a launch config 'Extension (Chrome)' to be able
to run the extension, but read the updated doc in browser.md first

Signed-off-by: nkomonen <nkomonen@amazon.com>

* run formatter

Signed-off-by: nkomonen <nkomonen@amazon.com>

---------

Signed-off-by: nkomonen <nkomonen@amazon.com>