fix: Use 'fs' methods if in C9
Problem:

We recently made some changes to our file system code
that used the vscode.workspace.fs instance. The issue
is that some of its methods are not supported in C9

Solution:

If in C9 use the original "fs" instance methods instead
for delete, readdir, and mkdir.

Also added tests.

Signed-off-by: nkomonen <nkomonen@amazon.com>