refactor: remove 'glob' from CW
Problem:
We cannot use 'glob' in browser mode and this breaks it

Solution:
Use vscode.workspace.findFiles.

Note the previous comments about it not working, which was due
to us needing to use a relative path since it searches the workspace
only. Additionally if there is no workspace it looks like it just searches
the project folder.

Signed-off-by: nkomonen <nkomonen@amazon.com>