feature(repo): generalize package script, remove redundant webpack.browser configs
package.json now accepts an argument to a subproject directory and will package the .vsix for that subproject.
This means that packages/amazonq can now be packaged into a .vsix with `npm run package`.
However, running this from the root project will only yield the toolkit.vsix for now.

Also, we can reference the browser config by name instead of file, so we don't need a separate config file for CLI usage.

Finally, fix an issue with the packaging script so that we always copy over the CHANGELOG.md, especially for releases.