feat: add PrepareCodeGenerationState to featureDev (#1803)
After a customer is satisfied with a suggested approach in /dev chat, it has the option to generate code by clicking Write Code.

This PR starts adding the functionality happening after that click happens. For that first is introduced the prepare code generation state which would upload the project's code again and set the state for code generation.