feat: add CodeGenerationState minimal logic for starting code generation. (#1806)
The first step after the preparation for code generation was done is to start the code generation. i.e. calling startTaskAssistCodeGeneration and handling its response. This logic is done through the CodeGenerationState. CodeGenerationState and PrepareCodeGenState are then looped while there is more feedback or work to be done in the codegeneeration.

In this PR is only the added the structure for code generation as explained before, in follow ups it will be added the polling mechanism for checking the status of code generation and finally the import of the new files. I'm leaving placeholder for that as well as for code generation follow ups.