feat(amazonq): handle invalidStateEvent response, show error #4448
Problem:
CodeWhispererStreaming has a new field in the ChatResponseStream payload which
is not currently handled in the toolkit. The field is used whenever the response
comes to an invalid state, and contains a message explaining the reasons for the
invalid state.

Solution:
Update the CodeWhispererStreaming generated client, so it includes this new field.
Add the correct handling of the invalid state event in the featureDev's client.