webpack: fix debugging issues
In a previous configuration of the webpack config we had offset
breakpoints that were fixed with 'eval-source-map' in the devtools field.

But later I found that this broke debugging the web unit tests.

Solution

As a solution, through trial and error, I found that using SourceMapDevToolPlugin
and excluding node_modules made the breakpoint offset issue disappear. So now
we can stop using 'eval-source-map' and everything seems to work.

Signed-off-by: nkomonen <nkomonen@amazon.com>