feat: add acceptCode handling (#1816)
After code is generated successfully the customer needs to chose between accepting code, i.e. saving the generated code changes, or providing feedback and regenerating code. In this case, it is being added the initial functionality of accepting code, which means saving new file changes, deleting files.

It leaves placeholders for the next possible interactions after accepting code, which are: starting a new task and closing the session. In addition, it leaves out code references as this will be implemented in follow ups.