feat: add closeSession and newTask handling. (#1819)
This is the last step in the /dev lifecycle after the customer has generated code and accepted the changes. This step let's him decide between finishing the chat, i.e. closing the session and blocking the current chat, or starting a new task, i.e. finishing the previous session and restarting in the same chat. After this, the process goes back to approach generation, thus restarting the /dev lifecycle.

Next steps: Adding telemetry for the code generation parts of the lifecycle and specially add unit tests. It will be also reviewed possible error handling situations and more.