fix: make getSession synchronized to avoid race conditions (#1836)
featureDev had a race condition where mynah ui would call in series two asynchronous functions that would try to create a new session and the map used as storage was creating two elements with the same key. To solve that is needed to use a thread safe function with the synchronize block.