ui(amazonq/codewhisperer): move tree nodes to cw status bar menu
As part of splitting the toolkit extension into an Amazon Q + CodeWhisperer extension, we are pre-emptively deprecating the CW tree view in favor of the status bar menu. The idea is that users will begin to get familiar with not having CW options in the toolkit explorer tree, and instead have them in the status bar menu that will be present in the standalone extension.

- Removed CodeWhisperer tree node
- Expanded the status bar menu with CodeWhisperer + AmazonQ + generic options'
- 'free tier limit reach' state was stored in the CW tree and only updated on refresh. It is now part of CW's "global" state. Some tree refreshes have been replaced with updating this global state.

Tested cases:
1. Not signed in
2. Signed in with builder ID
3. Signed in with IDC
4. Signed in with IDC + new customizations available