feat: add code reference handling to featureDev (#1831)
featureDev, a.k.a /dev in amazonq has two moments where references are handled. One, when it is shown in the diff view chat component, where all references are listed for the customer to see. Two, when the customer accepts code generation where these references are stored in the general chat session.

This PR is adding a storage for the reference logs and also sending the references to the chat view with the formatting that /dev expects.