feat: force refesh project files after accepting code (#1839)
JetBrains does not refresh files automatically after the file system is updated, which is expected. However, for the customer experience of featureDev it is needed to have the project files be refreshed after accepting code so the customer has feedback that they had been created successfully. It was decided to use the async refreshing to let the UI and JetBraisn continue why these files are being refreshed. In summary, it will look like the user decided to reload from disk.