fix(cwspr): update codePercentage metrics threshold from 500 to 50 #4522
Problem:
By conducting one week of data collection, we find that multi character text change which has more than 50 characters only account for 2% of total document change, but they somehow contributed to more than 98% of characters in the doc. This is because user rarely copy and paste code into their editor, but when they do, the copy will result in huge number of character increase. This kind of increase is not what we want to capture in the cwspr percentage code written metric.

50 is a good threshold that includes copy of small code snippets and IDE intelliSense accepted code.

Solution:
Ignore any multi character input more than 50 characters.
The userDetails JSON will be removed in the future. We want to monitor it for a few weeks longer.