auth: log errors when we go from valid to invalid state (#4553)
Problem:
When errors occurr somewhere in the getToken() process, if it
fails we do not log the actual underlying error anywhere. We simply
tell them to reauth.

This caused issues when debugging premature auth expiration problems since couldn't see why.

Solution:
If we switch from valid to invalid connection state, then log the error
as a warning. This should give us the reason as to why it is failing, though
we will see also see the expected case where it actually expired.

Signed-off-by: Nikolas Komonen <nkomonen@amazon.com>