auth: handle networking issues during sso validation
Problem:

We were having issues due to network issues,
such as no internet connected, the SSO auth token
would become invalid.

Solution:

We had some code that validated the token, but if it failed
due to network issues it would invalidate it.

Now, when we validate the token and there is a network error,
we will keep the state as-is instead of setting the connection
to invalid.

The user will instead see an error message indicating network issues.

Signed-off-by: Nikolas Komonen <nkomonen@amazon.com>