auth: add more network related errors to check
Problem:

We have a function isNetworkError that determines if a
an error is network related.

The issue is that we can see other network related errors
in our telemetry metric aws_refreshCredentials that are not
caught by isNetworkError.

Solution:

Add some of the network errors that we see in telemetry to
the isNetworkError check.

This will improve false token invalidations during the sso token
refresh process since we were previously invalidating incorrectly
on certain network issues

Signed-off-by: Nikolas Komonen <nkomonen@amazon.com>