telemetry: aws_loginWithBrowser includes user polling (#4605)
Problem:

Before, aws_loginWithBrowser metric only represented the opening
of the browser window and not the part where users would login and
confirm scopes. So if they failed somewhere in the browser,
aws_loginWithBrowser would still result in Succeeded.

We did not have good insight in to the success of the part where
we poll until the user completes the browser login.

Solution:

Now, aws_loginWithBrowser will include the part where the user must
login and accept scopes. Only once that is done do we emit Succeeded.

Signed-off-by: Nikolas Komonen <nkomonen@amazon.com>