test: fix core tests prompting build task
Problem:

When we run the core tests they prompt for a build task
before running. We should not have to do this.

Solution:

Create a specific build task and explicitly use that.

I think due to our recent changes the "defaultBuildTask" cannot
be appropriately resolved which is why this change was needed.

Signed-off-by: Nikolas Komonen <nkomonen@amazon.com>