refactor: remove index files in core
Problem:

We had index files in the core package which we used to expose
certain methods once we packaged core in to a node module
for use by the toolkit and amazonq package.

We don't need it.

Solution:

Stop using the index files and go back to using the extension*.ts files

Signed-off-by: Nikolas Komonen <nkomonen@amazon.com>