refactor: dynamically build webpack configs
Now we can dynamically build webpack configs by
exporting a function instead of the config.

This will allow us to check for the 'development' mode
at creation and from there we can modify the webpack
config that we return.

So now if we have `webpack --mode development`, we can
recognize we are in development mode and incrementally change
our config to our liking.

Signed-off-by: Nikolas Komonen <nkomonen@amazon.com>