fix(amazon-q featureDev): upload root dir instead of /src dir #4654
Problem
Currently, we're looking for the /src folder to upload, and if not found we're falling back to uploading the root directory. this was causing to missing multiple relevant files (e.g test directories) for the code generation.

Solution
We're changing the Amazon Q feature dev to upload workspaces' root directory instead of looking for /src folder so that we'll include all relevant files during plan and code generation.