refactor: browser test output unique file
Problem:

We need to both compile all source code + webpack a web extension
when running the Web unit tests so that we do not have type errors
+ have an executable file (webpacked file)

The problem is that the name extensionWeb.js is shared by both the
compiled output AND webpacked output. So one gets overwritten.

Solution:

Change the name of the webpacked output so that it does not get
overwritten. Now in unit tests we target that specifically.

Signed-off-by: Nikolas Komonen <nkomonen@amazon.com>