feat(amazonqFeatureDev): add partial code acceptance (#4244)
Mynah UI already supports an option to handle file clicking. This is being leveraged to add a partial code acceptance feautre to the /dev chat. That is, the user will be able to click through the file changes generated from code generation step to reject them or not. These changes are similar to the ones already implemented for the VSCode toolkit.

---------

Co-authored-by: Nikita Mikhailov <noobgam@amazon.de>