fix(amazonq): "aws.amazonq.refreshConnectionCallback" not found #4674
Problem:
When Q activates, if at that time toolkit is installed but not activated, Q will throw aws.amazonq.refreshConnectionCallback' cmd not found. Even though this is caught in a tryExecute, it will generate noise for debugging.

This edge case can happen when reloading webview and both extension activates, Q activates first.

Solution:
Avoid the throw by adding if-else to make sure grabbing the toolkit API must happen with a activated toolkit.