refactor: web mode webpack
Before we had a custom flag to not build certain Web mode
bundles. But it is instead easier to always build it, but output
the bundle with a different name.

We did this due to overlapping output files with the same name,
but with this new change it does not happen anymore.

Signed-off-by: Nikolas Komonen <nkomonen@amazon.com>