amazonq: right click context menu for current line
Problem:

If the user has not highlighted any text in the editor, a right click
will not show anything in the context menu. But if they select some
text when right click it will show the Amazon Q Chat context commands.

Solution:

If the user right clicks and there is no highlight, show the context menu
and it will automatically use the current line

Signed-off-by: Nikolas Komonen <nkomonen@amazon.com>