amazonq: codelens to prompt user to try chat
For a set amount of times we will show the user a codelens
above the currently selected line that prompts them to try
and use Q chat.
We are doing this as a way to advertise the chat feature.

The codelens will show for a specific amount of times and then
never show again, so we do not annoy users.

Clicking the codelens opens Q chat

Signed-off-by: Nikolas Komonen <nkomonen@amazon.com>