fix(amazonqFeatureDev): race conditions on session creation #4724
Problem:
When a customer starts a /dev chat with a message, e.g. /dev <do something>, the
connector logic calls two event handlers in amazonqFeatureDev logic parallely
that try to create a session, this creates a race condition where the customer
could get to code generation with the wrong session, i.e. one with no
conversationId.

Solution:
Lock the getSession function to mitigate concurrent calls.

This is a similar approach to the one taken in JetBrain's toolkit:
https://github.com/aws/aws-toolkit-jetbrains/blob/ec168fc6e9306f22f07945fd516d6acdc7fe396e/plugins/amazonq/chat/jetbrains-community/src/software/aws/toolkits/jetbrains/services/amazonqFeatureDev/storage/ChatSessionStorage.kt#L18