fix(globals): use 'local' globals if not running web mode (#4754)
Problem: To get web tests working we used globalThis for globals. This is shared across extensions in VS Code, resulting in problems if another extension declares the same global twice. This is occurs when we run standalone and toolkit at the same time.

Solution: Use local globals if we are not in web mode. Otherwise, use full globals. This allows the web tests to pass. We will have to revisit this if we have 2 extensions running in web mode.

- Also add slightly more informative globals error handling.
- Refactor CW class so that globals doesn't get called during imports.