Merge master into feature/standalone (#4753)
* Release 2.20.0

* Update version to snapshot version: 2.21.0-SNAPSHOT

* amazonq: vscode walkthrough (#4739)

- Creates the Amazon Q walkthrough
- GIFs still need to be created, we are waiting on those but we can merge for now
- When the user logs in to Amazon Q the first time they will be shown the walkthrough page. It will never show again
- We have telemetry for when the walkthrough is shown through our "first login" automation, and when certain events are triggered in the walkthrough such as opening the chat

Signed-off-by: Nikolas Komonen <nkomonen@amazon.com>

* fix(globals): use 'local' globals if not running web mode (#4754)

Problem: To get web tests working we used globalThis for globals. This is shared across extensions in VS Code, resulting in problems if another extension declares the same global twice. This is occurs when we run standalone and toolkit at the same time.

Solution: Use local globals if we are not in web mode. Otherwise, use full globals. This allows the web tests to pass. We will have to revisit this if we have 2 extensions running in web mode.

- Also add slightly more informative globals error handling.
- Refactor CW class so that globals doesn't get called during imports.

* Release 0.12.0

* Update version to snapshot version: 0.13.0-SNAPSHOT

* Release 0.13.0

* Update version to snapshot version: 0.14.0-SNAPSHOT

* Release 0.14.0

* Update version to snapshot version: 0.15.0-SNAPSHOT

* telemetry(amazonq chat): params for Q chat #4714

Problem:
- Missing parameters in `onCopyCodeToClipboard` and `onCodeInsertToCursorPosition` events for telemetry.

Solution:
- Added `cwsprChatCodeBlockIndex` and `cwsprChatTotalCodeBlocks` parameters to the `onCopyCodeToClipboard` and `onCodeInsertToCursorPosition` events.
- `cwsprChatCodeBlockIndex` - Provides the index of that code block, when a user interacts with a code block within a message in the conversation, 
- `cwsprChatTotalCodeBlocks` - Total number of code blocks inside a generated message.
- Passed these two parameters to `insert_code_at_cursor_position` and `code_was_copied_to_clipboard` telemetry events.

* feat: beautify auth page after pkce login (#4738)

* feat: beautify auth page after pkce login

Problem:
- Old screen was ugly and basic

Solution:
- New screen is similiar to the current oidc login webpage

NOTE: We need to manually copy the html to dist because the vue loader doesn't do it for us. Also, it's a lot simpler to just serve the html rather than do server side loading for the vue component

* amazonq: command to show the chat window (#4756)

Before, we had a function to show the chat webview or used the
vscode generated show chat command.

Now we have a single command which can be used anywhere. This will
allow us to get better metrics on when the chat window is opened.

Signed-off-by: Nikolas Komonen <nkomonen@amazon.com>

* test: fik flaky fs test (#4764)

Our mkdir() spy was sometimes catching an unexpected mkdir()
call and causing the test to fail.

So instead we are just checking that the first mkdir() call
has the expected args

Signed-off-by: Nikolas Komonen <nkomonen@amazon.com>

* fix(amazonq): remove Welcome Page #4736

Remove onboarding page in favor of walkthrough #4739

* feat(amazonq): marketplace icon #4767

Update Q icon. This png is currently only used for the marketplace listing.

---------

Signed-off-by: Nikolas Komonen <nkomonen@amazon.com>
Co-authored-by: aws-toolkit-automation <>
Co-authored-by: Nikolas Komonen <118216176+nkomonen-amazon@users.noreply.github.com>
Co-authored-by: Maxim Hayes <149123719+hayemaxi@users.noreply.github.com>
Co-authored-by: Laxman Reddy <141967714+laileni-aws@users.noreply.github.com>
Co-authored-by: Josh Pinkney <103940141+jpinkney-aws@users.noreply.github.com>
Co-authored-by: Vikash Agrawal <vikashagrawal1990@gmail.com>
Co-authored-by: Justin M. Keyes <jmkeyes@amazon.com>
Co-authored-by: Nikolas Komonen <nkomonen@amazon.com>